+++ changelog.md
@@ -17,6 +17,7 @@ Codion Change Log
 - EntityQueryModel.referenceDepth() configuration added.
 - DefaultEntityEditor.setDetail() and replaceDetail() refactored.
 - EntityEditor.DetailEditors.remove() added.
+- EntitySearchModel.Builder.searchColumns() overloaded with varargs parameter.
 ### is.codion.swing.framework.ui
 - EditorComponents.entries() and editorComponent() bug fixed, now exclude entries without a component.
 - EditorComponents.detail() added, removal implemented.